testing jersey implementations
