--- src/main/java/annotations/mongo/documents/DocumentEntryKeys.java
@@ -1,12 +1,0 @@
-package annotations.mongo.documents;
-
-import static java.lang.annotation.ElementType.TYPE;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-@Retention(RUNTIME)
-@Target(TYPE)
-public @interface DocumentEntryKeys {
-}
