+++ src/main/java/com/the/mild/project/db/mongo/annotations/DocumentEntryKeys.java
@@ -1,0 +1,12 @@
+package com.the.mild.project.db.mongo.annotations;
+
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+@Retention(RUNTIME)
+@Target(TYPE)
+public @interface DocumentEntryKeys {
+}
