+++ src/main/java/com/the/mild/project/db/mongo/annotations/DocumentSerializable.java
@@ -1,0 +1,13 @@
+package com.the.mild.project.db.mongo.annotations;
+
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+@Retention(RUNTIME)
+@Target(TYPE)
+public @interface DocumentSerializable {
+    String collectionName();
+}
