+++ src/main/java/com/the/mild/project/db/mongo/documents/util/BaseDocument.java
@@ -1,0 +1,12 @@
+package com.the.mild.project.db.mongo.documents.util;
+
+import org.bson.Document;
+
+public interface BaseDocument {
+
+    /**
+     * Get back a document
+     * @return document
+     */
+    Document getDocument();
+}
