+++ src/main/java/com/the/mild/project/db/mongo/documents/util/InsertDocumentEntry.java
@@ -1,0 +1,5 @@
+package com.the.mild.project.db.mongo.documents.util;
+
+public interface InsertDocumentEntry {
+    String key();
+}
