+++ src/main/java/com/the/mild/project/db/mongo/util/exceptions/CollectionNotFoundException.java
@@ -1,0 +1,7 @@
+package com.the.mild.project.db.mongo.util.exceptions;
+
+public class CollectionNotFoundException extends Exception {
+    public CollectionNotFoundException(String message) {
+        super(message);
+    }
+}
