+++ src/main/java/com/the/mild/project/db/mongo/util/exceptions/DocumentSerializationException.java
@@ -1,0 +1,7 @@
+package com.the.mild.project.db.mongo.util.exceptions;
+
+public class DocumentSerializationException extends Exception {
+    public DocumentSerializationException(String errorMessage) {
+        super(errorMessage);
+    }
+}
