+++ src/main/java/com/the/mild/project/server/resources/TestResource.java
@@ -1,0 +1,43 @@
+package com.the.mild.project.server.resources;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.MediaType;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
+import com.the.mild.project.server.jackson.JacksonTest;
+
+/**
+ * Root resource (exposed at "myresource" path)
+ */
+@Path("testresource")
+public class TestResource {
+
+    /**
+     * Method handling HTTP GET requests. The returned object will be sent
+     * to the client as "text/plain" media type.
+     *
+     * @return String that will be returned as a text/plain response.
+     */
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    public String getIt() {
+        final ObjectMapper mapper = new ObjectMapper();
+
+        final JacksonTest test = new JacksonTest("test", "resource");
+
+        String result = "{}";
+        try {
+            result = mapper.writeValueAsString(test);
+            System.out.println(result);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+        }
+        return result;
+    }
+}
