--- src/main/java/db/mongo/documents/util/BaseDocument.java
@@ -1,12 +1,0 @@
-package db.mongo.documents.util;
-
-import org.bson.Document;
-
-public interface BaseDocument {
-
-    /**
-     * Get back a document
-     * @return document
-     */
-    Document getDocument();
-}
