--- src/main/java/db/mongo/documents/util/InsertDocumentEntry.java
@@ -1,5 +1,0 @@
-package db.mongo.documents.util;
-
-public interface InsertDocumentEntry {
-    String key();
-}
