--- src/main/java/db/mongo/util/exceptions/CollectionNotFoundException.java
@@ -1,7 +1,0 @@
-package db.mongo.util.exceptions;
-
-public class CollectionNotFoundException extends Exception {
-    public CollectionNotFoundException(String message) {
-        super(message);
-    }
-}
