--- src/main/java/db/mongo/util/exceptions/DocumentSerializationException.java
@@ -1,7 +1,0 @@
-package db.mongo.util.exceptions;
-
-public class DocumentSerializationException extends Exception {
-    public DocumentSerializationException(String errorMessage) {
-        super(errorMessage);
-    }
-}
