+++ src/test/java/mongo/DocumentUT.java
@@ -1,13 +1,13 @@
 package mongo;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.Assert.assertTrue;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
 public class DocumentUT {
 
     @Test
     public void validTestDocument() {
-        assertTrue(true, "test");
+        assertTrue("test", true);
     }
 }
